Now, the package is installed simply with Pip.

+++ pointnet2_ops/_ext-src/include/group_points.h
@@ -2,4 +2,4 @@
 #include <torch/extension.h>
 
 at::Tensor group_points(at::Tensor points, at::Tensor idx);
-at::Tensor group_points_grad(at::Tensor grad_out, at::Tensor idx, const int n);
+at::Tensor group_points_grad(at::Tensor grad_out, at::Tensor idx, const int64_t n);
